fix(tests): reduce sonar architecture boundary violations
